python/gcp-secrect-manager/pyproject.toml
[tool.poetry]
name = "gcp-secrect-manager"
version = "0.1.0"
description = ""
authors = ["perebaj <perebaj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
google-cloud-secret-manager = "^2.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
